Stable working version using our version of robotApi (not working yet)

+++ TelemedicinaMinibot/app/src/main/java/com/intec/telemedicina/CoordinatePose.kt
@@ -3,7 +3,7 @@ package com.intec.telemedicina
 import com.ainirobot.coreservice.client.actionbean.Pose
 
 data class CoordinatePose(
-    val x : Float,
-    val y : Float,
-    val theta : Float
+    val x : Double,
+    val y : Double,
+    val theta : Double
 )
